refactor(settings-dialog): extract tab bodies into a part file
Move the General / Appearance / Emulators tab builders, the system-image &
logo pickers, and the shared switch/preview helpers out of
system_emulator_settings_dialog.dart into a new part file
`system_emulator_settings_dialog/tabs.dart` as `extension _Tabs`.

Pure move, zero behaviour change. State stays on the host State; `setState`
calls route through the host `rebuild` bridge and the static `_log` is
referenced qualified. Host 1534 → 894 — closes out the settings-dialog split.
